add controller

+++ src/main/java/com/example/librarymanagementbackend/dto/base/request/BaseGetAllRequest.java
@@ -6,6 +6,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public abstract class BaseGetAllRequest {
-    protected Integer skipCount = 0;
-    protected Integer maxResultCount = 10;
+    protected Long skipCount = 0L;
+    protected Long maxResultCount = 10L;
 }
