+++ src/main/java/com/example/librarymanagementbackend/dto/user/request/UserCreationRequest.java
@@ -16,5 +16,5 @@ public class UserCreationRequest {
     String email;
     @Size(min = 8, message = "Password must be at least 8 characters")
     String password;
-    int roleId;
+    Long roleId;
 }
